requirements.txt
numpy==2.0.1
opencv-python==4.10.0.84
paho-mqtt==2.1.0
pillow==10.4.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyzbar==0.1.9
